handson/h02-ec2-dnn/requirements.txt
-e .
autopep8>=1.5.4
aws-cdk.aws-ec2>=1.30.0
